Merge branch 'CPG130_IssueFixes_Anand' into 'CPG133_Release'

See merge request IvyProjects/IvyCpg_Product_MobileApp_Android!786

(cherry picked from commit 509bc440e19fb79b08dbe037470f1a414ee8df4c)

3d3ab2b3 Merge branch 'CPG130_IssueFixes_Anand' into 'CPG130_Release'

+++ ivycpg/src/main/java/com/ivy/sd/png/provider/PrintHelper.java
@@ -2812,12 +2812,14 @@ public class PrintHelper {
             File folder = new File(path);
 
             File sfFiles[] = folder.listFiles();
-            for (int i = 0; i < sfFiles.length; i++) {
-                File deleteFile = new File(folder, "/" + sfFiles[i].getName());
-                deleteFile.delete();
+            if (sfFiles != null) {
+                for (int i = 0; i < sfFiles.length; i++) {
+                    File deleteFile = new File(folder, "/" + sfFiles[i].getName());
+                    deleteFile.delete();
 
+                }
+                folder.delete();
             }
-            folder.delete();
         } catch (Exception e) {
             Commons.printException(e);
         }
